+++ Sources/TuistGraph/DependenciesGraph/ThirdPartyDependency.swift
@@ -4,7 +4,7 @@ import TSCBasic
 // A enum containing information about third party dependency.
 public enum ThirdPartyDependency: Hashable, Equatable, Codable {
     /// A dependency that represents a pre-compiled .xcframework.
-    case xcframework(path: AbsolutePath, architectures: Set<BinaryArchitecture>)
+    case xcframework(path: AbsolutePath)
 }
 
 // MARK: - Codable
@@ -17,7 +17,6 @@ extension ThirdPartyDependency {
     private enum CodingKeys: String, CodingKey {
         case kind
         case path
-        case architectures
     }
 
     public init(from decoder: Decoder) throws {
@@ -26,18 +25,16 @@ extension ThirdPartyDependency {
         switch kind {
         case .xcframework:
             let path = try container.decode(AbsolutePath.self, forKey: .path)
-            let architectures = try container.decode(Set<BinaryArchitecture>.self, forKey: .architectures)
-            self = .xcframework(path: path, architectures: architectures)
+            self = .xcframework(path: path)
         }
     }
 
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         switch self {
-        case let .xcframework(path, architectures):
+        case let .xcframework(path):
             try container.encode(Kind.xcframework, forKey: .kind)
             try container.encode(path, forKey: .path)
-            try container.encode(architectures, forKey: .architectures)
         }
     }
 }