+++ Tests/TuistDependenciesTests/Carthage/Models/CarthageVersionFileTests.swift
@@ -28,6 +28,15 @@ final class CarthageVersionFileTests: TuistUnitTestCase {
         XCTAssertDecodableEqualToJson(json, expected)
     }
 
+    func test_codable_realmSwift() {
+        // Given
+        let json = CarthageVersionFile.testRealmCocoaJson
+        let expected = CarthageVersionFile.testRealmCocoa
+
+        // When / Then
+        XCTAssertDecodableEqualToJson(json, expected)
+    }
+
     func test_allProducts() {
         // Given
         let iOSProduct = CarthageVersionFile.Product.test(name: "iOS")
@@ -47,48 +56,6 @@ final class CarthageVersionFileTests: TuistUnitTestCase {
 
         // Then
         let expected: [CarthageVersionFile.Product] = [iOSProduct, macOSProduct, watchOSProduct, tvOSProduct]
-        XCTAssertEqual(got, expected)
-    }
-
-    func test_product_architectures_arm64_x8664() {
-        // Given
-        let subject: CarthageVersionFile.Product = .test(
-            identifier: "macos-arm64_x86_64"
-        )
-
-        // When
-        let got = subject.architectures
-
-        // Then
-        let expected: [BinaryArchitecture] = [.arm64, .x8664]
-        XCTAssertEqual(got, expected)
-    }
-
-    func test_product_architectures_arm64_i386_x8664() {
-        // Given
-        let subject: CarthageVersionFile.Product = .test(
-            identifier: "ios-arm64_i386_x86_64-simulator"
-        )
-
-        // When
-        let got = subject.architectures
-
-        // Then
-        let expected: [BinaryArchitecture] = [.arm64, .i386, .x8664]
-        XCTAssertEqual(got, expected)
-    }
-
-    func test_product_architectures_arm6432_armv7k() {
-        // Given
-        let subject: CarthageVersionFile.Product = .test(
-            identifier: "watchos-arm64_32_armv7k"
-        )
-
-        // When
-        let got = subject.architectures
-
-        // Then
-        let expected: [BinaryArchitecture] = [.arm6432, .armv7k]
         XCTAssertEqual(got, expected)
     }
 }