+++ Tests/TuistGraphTests/DependenciesGraph/ThirdPartyDependencyTests.swift
@@ -7,9 +7,7 @@ import XCTest
 final class ThirdPartyDependencyTests: TuistUnitTestCase {
     func test_codable_xcframework() {
         // Given
-        let subject = ThirdPartyDependency.testXCFramework(
-            architectures: [.arm64, .i386, .arm6432]
-        )
+        let subject = ThirdPartyDependency.testXCFramework()
 
         // Then
         XCTAssertCodable(subject)